+++ NEWS.md
@@ -1,6 +1,8 @@
 # phenoptrReports 0.2.6.9000
 
 Analysis addin:
+- The Count Within worksheet includes summary counts across all tissue 
+  categories when multiple categories are reported.
 - Fix chart report to run when 'Total Cells' is the only 'phenotype' 
   defined (#32).
 - Fix H-Score report to always include every combination of